requirements.txt
ultralytics>=8.0.196
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.23.0
pandas>=2.0.0
tqdm>=4.65.0
pyyaml>=6.0
cvxpy>=1.4.2
ecos>=2.0.13
osqp>=0.6.5
scipy>=1.10.0
